fix: resolve review issues — DK-2, dead code, PROC_MAX, warnings
- .gitignore: add engine/canvasos and tests/ binaries (no more accidental binary commits)
- engine.c: wrap Phase 1 demo statics (canvas, aset, tick, run_frame_adaptive) in #ifdef ENGINE_DEMO_MAIN — eliminates 'defined but not used' warnings
- canvasos_types.h: replace float density + activeset_mode(float) with integer open_count threshold (DK-2 compliance; thresholds 82/1229 of TILE_COUNT=4096)
- active_set.c: remove all float density calculations, call activeset_mode(open_count)
- test_gate.c: update density printf to use integer arithmetic
- scheduler.c: fix sched_tick indentation — sc->ctx WH record block now correctly nested inside if(energy==0)
- canvasos_sched.h: PROC_MAX 64→128 (bottleneck relief; dynamic allocation planned Phase 8)
- canvas_multiverse.h: lane_count uint8_t→uint16_t (LANE_ID_MAX=256 comparison now valid)
- canvas_determinism.h: suppress 'label' unused-parameter warning in release builds
- canvas_bh_compress.c: suppress 'guard' unused-parameter warning when ASSERT_TICK_BOUNDARY is no-op

Build: 0 errors, 0 warnings | Tests: all PASS (test_scan, test_gate, test_canvasfs x6, test_scheduler, test_cvp x7)

+++ engine/include/canvas_multiverse.h
@@ -91,8 +91,8 @@ static inline WH_BH_Zone zone_q2(void) {
 
 /* ---- Multiverse 설정 ---- */
 typedef struct {
-    /* 활성 Lane 수 */
-    uint8_t      lane_count;
+    /* 활성 Lane 수 (uint16_t: LANE_ID_MAX=256 비교 가능) */
+    uint16_t     lane_count;
     LaneTable    lanes;
 
     /* 활성 Branch 수 */
